Split out theme code to more files
