+++ theme/theme.go
@@ -10,224 +10,6 @@ import (
 )
 
 const (
-	// ColorRed is the red primary color name.
-	//
-	// Since: 1.4
-	ColorRed = "red"
-	// ColorOrange is the orange primary color name.
-	//
-	// Since: 1.4
-	ColorOrange = "orange"
-	// ColorYellow is the yellow primary color name.
-	//
-	// Since: 1.4
-	ColorYellow = "yellow"
-	// ColorGreen is the green primary color name.
-	//
-	// Since: 1.4
-	ColorGreen = "green"
-	// ColorBlue is the blue primary color name.
-	//
-	// Since: 1.4
-	ColorBlue = "blue"
-	// ColorPurple is the purple primary color name.
-	//
-	// Since: 1.4
-	ColorPurple = "purple"
-	// ColorBrown is the brown primary color name.
-	//
-	// Since: 1.4
-	ColorBrown = "brown"
-	// ColorGray is the gray primary color name.
-	//
-	// Since: 1.4
-	ColorGray = "gray"
-
-	// ColorNameBackground is the name of theme lookup for background color.
-	//
-	// Since: 2.0
-	ColorNameBackground fyne.ThemeColorName = "background"
-
-	// ColorNameButton is the name of theme lookup for button color.
-	//
-	// Since: 2.0
-	ColorNameButton fyne.ThemeColorName = "button"
-
-	// ColorNameDisabledButton is the name of theme lookup for disabled button color.
-	//
-	// Since: 2.0
-	ColorNameDisabledButton fyne.ThemeColorName = "disabledButton"
-
-	// ColorNameDisabled is the name of theme lookup for disabled foreground color.
-	//
-	// Since: 2.0
-	ColorNameDisabled fyne.ThemeColorName = "disabled"
-
-	// ColorNameError is the name of theme lookup for foreground error color.
-	//
-	// Since: 2.0
-	ColorNameError fyne.ThemeColorName = "error"
-
-	// ColorNameFocus is the name of theme lookup for focus color.
-	//
-	// Since: 2.0
-	ColorNameFocus fyne.ThemeColorName = "focus"
-
-	// ColorNameForeground is the name of theme lookup for foreground color.
-	//
-	// Since: 2.0
-	ColorNameForeground fyne.ThemeColorName = "foreground"
-
-	// ColorNameHeaderBackground is the name of theme lookup for background color of a collection header.
-	//
-	// Since: 2.4
-	ColorNameHeaderBackground fyne.ThemeColorName = "headerBackground"
-
-	// ColorNameHover is the name of theme lookup for hover color.
-	//
-	// Since: 2.0
-	ColorNameHover fyne.ThemeColorName = "hover"
-
-	// ColorNameHyperlink is the name of theme lookup for hyperlink color.
-	//
-	// Since: 2.4
-	ColorNameHyperlink fyne.ThemeColorName = "hyperlink"
-
-	// ColorNameInputBackground is the name of theme lookup for background color of an input field.
-	//
-	// Since: 2.0
-	ColorNameInputBackground fyne.ThemeColorName = "inputBackground"
-
-	// ColorNameInputBorder is the name of theme lookup for border color of an input field.
-	//
-	// Since: 2.3
-	ColorNameInputBorder fyne.ThemeColorName = "inputBorder"
-
-	// ColorNameMenuBackground is the name of theme lookup for background color of menus.
-	//
-	// Since: 2.3
-	ColorNameMenuBackground fyne.ThemeColorName = "menuBackground"
-
-	// ColorNameOverlayBackground is the name of theme lookup for background color of overlays like dialogs.
-	//
-	// Since: 2.3
-	ColorNameOverlayBackground fyne.ThemeColorName = "overlayBackground"
-
-	// ColorNamePlaceHolder is the name of theme lookup for placeholder text color.
-	//
-	// Since: 2.0
-	ColorNamePlaceHolder fyne.ThemeColorName = "placeholder"
-
-	// ColorNamePressed is the name of theme lookup for the tap overlay color.
-	//
-	// Since: 2.0
-	ColorNamePressed fyne.ThemeColorName = "pressed"
-
-	// ColorNamePrimary is the name of theme lookup for primary color.
-	//
-	// Since: 2.0
-	ColorNamePrimary fyne.ThemeColorName = "primary"
-
-	// ColorNameScrollBar is the name of theme lookup for scrollbar color.
-	//
-	// Since: 2.0
-	ColorNameScrollBar fyne.ThemeColorName = "scrollBar"
-
-	// ColorNameSelection is the name of theme lookup for selection color.
-	//
-	// Since: 2.1
-	ColorNameSelection fyne.ThemeColorName = "selection"
-
-	// ColorNameSeparator is the name of theme lookup for separator bars.
-	//
-	// Since: 2.3
-	ColorNameSeparator fyne.ThemeColorName = "separator"
-
-	// ColorNameShadow is the name of theme lookup for shadow color.
-	//
-	// Since: 2.0
-	ColorNameShadow fyne.ThemeColorName = "shadow"
-
-	// ColorNameSuccess is the name of theme lookup for foreground success color.
-	//
-	// Since: 2.3
-	ColorNameSuccess fyne.ThemeColorName = "success"
-
-	// ColorNameWarning is the name of theme lookup for foreground warning color.
-	//
-	// Since: 2.3
-	ColorNameWarning fyne.ThemeColorName = "warning"
-
-	// SizeNameCaptionText is the name of theme lookup for helper text size, normally smaller than regular text size.
-	//
-	// Since: 2.0
-	SizeNameCaptionText fyne.ThemeSizeName = "helperText"
-
-	// SizeNameInlineIcon is the name of theme lookup for inline icons size.
-	//
-	// Since: 2.0
-	SizeNameInlineIcon fyne.ThemeSizeName = "iconInline"
-
-	// SizeNameInnerPadding is the name of theme lookup for internal widget padding size.
-	//
-	// Since: 2.3
-	SizeNameInnerPadding fyne.ThemeSizeName = "innerPadding"
-
-	// SizeNameLineSpacing is the name of theme lookup for between text line spacing.
-	//
-	// Since: 2.3
-	SizeNameLineSpacing fyne.ThemeSizeName = "lineSpacing"
-
-	// SizeNamePadding is the name of theme lookup for padding size.
-	//
-	// Since: 2.0
-	SizeNamePadding fyne.ThemeSizeName = "padding"
-
-	// SizeNameScrollBar is the name of theme lookup for the scrollbar size.
-	//
-	// Since: 2.0
-	SizeNameScrollBar fyne.ThemeSizeName = "scrollBar"
-
-	// SizeNameScrollBarSmall is the name of theme lookup for the shrunk scrollbar size.
-	//
-	// Since: 2.0
-	SizeNameScrollBarSmall fyne.ThemeSizeName = "scrollBarSmall"
-
-	// SizeNameSeparatorThickness is the name of theme lookup for the thickness of a separator.
-	//
-	// Since: 2.0
-	SizeNameSeparatorThickness fyne.ThemeSizeName = "separator"
-
-	// SizeNameText is the name of theme lookup for text size.
-	//
-	// Since: 2.0
-	SizeNameText fyne.ThemeSizeName = "text"
-
-	// SizeNameHeadingText is the name of theme lookup for text size of a heading.
-	//
-	// Since: 2.1
-	SizeNameHeadingText fyne.ThemeSizeName = "headingText"
-
-	// SizeNameSubHeadingText is the name of theme lookup for text size of a sub-heading.
-	//
-	// Since: 2.1
-	SizeNameSubHeadingText fyne.ThemeSizeName = "subHeadingText"
-
-	// SizeNameInputBorder is the name of theme lookup for input border size.
-	//
-	// Since: 2.0
-	SizeNameInputBorder fyne.ThemeSizeName = "inputBorder"
-
-	// SizeNameInputRadius is the name of theme lookup for input corner radius.
-	//
-	// Since: 2.4
-	SizeNameInputRadius fyne.ThemeSizeName = "inputRadius"
-
-	// SizeNameSelectionRadius is the name of theme lookup for selection corner radius.
-	//
-	// Since: 2.4
-	SizeNameSelectionRadius fyne.ThemeSizeName = "selectionRadius"
-
 	// VariantDark is the version of a theme that satisfies a user preference for a light look.
 	//
 	// Since: 2.0
@@ -242,21 +24,6 @@ const (
 	variantNameUserPreference fyne.ThemeVariant = 2 // locally used in builtinTheme for backward compatibility
 )
 
-// BackgroundColor returns the theme's background color.
-func BackgroundColor() color.Color {
-	return safeColorLookup(ColorNameBackground, currentVariant())
-}
-
-// ButtonColor returns the theme's standard button color.
-func ButtonColor() color.Color {
-	return safeColorLookup(ColorNameButton, currentVariant())
-}
-
-// CaptionTextSize returns the size for caption text.
-func CaptionTextSize() float32 {
-	return current().Size(SizeNameCaptionText)
-}
-
 // DarkTheme defines the built-in dark theme colors and sizes.
 //
 // Deprecated: This method ignores user preference and should not be used, it will be removed in v3.0.
@@ -265,46 +32,6 @@ func DarkTheme() fyne.Theme {
 
 	theme.initFonts()
 	return theme
-}
-
-// DefaultEmojiFont returns the font resource for the built-in emoji font.
-// This may return nil if the application was packaged without an emoji font.
-//
-// Since: 2.4
-func DefaultEmojiFont() fyne.Resource {
-	return emoji
-}
-
-// DefaultTextBoldFont returns the font resource for the built-in bold font style.
-func DefaultTextBoldFont() fyne.Resource {
-	return bold
-}
-
-// DefaultTextBoldItalicFont returns the font resource for the built-in bold and italic font style.
-func DefaultTextBoldItalicFont() fyne.Resource {
-	return bolditalic
-}
-
-// DefaultTextFont returns the font resource for the built-in regular font style.
-func DefaultTextFont() fyne.Resource {
-	return regular
-}
-
-// DefaultTextItalicFont returns the font resource for the built-in italic font style.
-func DefaultTextItalicFont() fyne.Resource {
-	return italic
-}
-
-// DefaultTextMonospaceFont returns the font resource for the built-in monospace font face.
-func DefaultTextMonospaceFont() fyne.Resource {
-	return monospace
-}
-
-// DefaultSymbolFont returns the font resource for the built-in symbol font.
-//
-// Since: 2.2
-func DefaultSymbolFont() fyne.Resource {
-	return symbol
 }
 
 // DefaultTheme returns a built-in theme that can adapt to the user preference of light or dark colors.
@@ -318,99 +45,6 @@ func DefaultTheme() fyne.Theme {
 	return defaultTheme
 }
 
-// DisabledButtonColor returns the theme's disabled button color.
-func DisabledButtonColor() color.Color {
-	return safeColorLookup(ColorNameDisabledButton, currentVariant())
-}
-
-// DisabledColor returns the foreground color for a disabled UI element.
-//
-// Since: 2.0
-func DisabledColor() color.Color {
-	return safeColorLookup(ColorNameDisabled, currentVariant())
-}
-
-// DisabledTextColor returns the theme's disabled text color - this is actually the disabled color since 1.4.
-//
-// Deprecated: Use theme.DisabledColor() colour instead.
-func DisabledTextColor() color.Color {
-	return DisabledColor()
-}
-
-// ErrorColor returns the theme's error foreground color.
-//
-// Since: 2.0
-func ErrorColor() color.Color {
-	return safeColorLookup(ColorNameError, currentVariant())
-}
-
-// FocusColor returns the color used to highlight a focused widget.
-func FocusColor() color.Color {
-	return safeColorLookup(ColorNameFocus, currentVariant())
-}
-
-// ForegroundColor returns the theme's standard foreground color for text and icons.
-//
-// Since: 2.0
-func ForegroundColor() color.Color {
-	return safeColorLookup(ColorNameForeground, currentVariant())
-}
-
-// HeaderBackgroundColor returns the color used to draw underneath collection headers.
-//
-// Since: 2.4
-func HeaderBackgroundColor() color.Color {
-	return current().Color(ColorNameHeaderBackground, currentVariant())
-}
-
-// HoverColor returns the color used to highlight interactive elements currently under a cursor.
-func HoverColor() color.Color {
-	return safeColorLookup(ColorNameHover, currentVariant())
-}
-
-// HyperlinkColor returns the color used for the Hyperlink widget and hyperlink text elements.
-func HyperlinkColor() color.Color {
-	return safeColorLookup(ColorNameHyperlink, currentVariant())
-}
-
-// IconInlineSize is the standard size of icons which appear within buttons, labels etc.
-func IconInlineSize() float32 {
-	return current().Size(SizeNameInlineIcon)
-}
-
-// InnerPadding is the standard gap between element content and the outside edge of a widget.
-//
-// Since: 2.3
-func InnerPadding() float32 {
-	return current().Size(SizeNameInnerPadding)
-}
-
-// InputBackgroundColor returns the color used to draw underneath input elements.
-func InputBackgroundColor() color.Color {
-	return current().Color(ColorNameInputBackground, currentVariant())
-}
-
-// InputBorderColor returns the color used to draw underneath input elements.
-//
-// Since: 2.3
-func InputBorderColor() color.Color {
-	return current().Color(ColorNameInputBorder, currentVariant())
-}
-
-// InputBorderSize returns the input border size (or underline size for an entry).
-//
-// Since: 2.0
-func InputBorderSize() float32 {
-	return current().Size(SizeNameInputBorder)
-}
-
-// InputRadiusSize returns the input radius size.
-//
-// Since: 2.4
-func InputRadiusSize() float32 {
-	return current().Size(SizeNameInputRadius)
-}
-
 // LightTheme defines the built-in light theme colors and sizes.
 //
 // Deprecated: This method ignores user preference and should not be used, it will be removed in v3.0.
@@ -421,189 +55,8 @@ func LightTheme() fyne.Theme {
 	return theme
 }
 
-// LineSpacing is the default gap between multiple lines of text.
-//
-// Since: 2.3
-func LineSpacing() float32 {
-	return current().Size(SizeNameLineSpacing)
-}
-
-// MenuBackgroundColor returns the theme's background color for menus.
-//
-// Since: 2.3
-func MenuBackgroundColor() color.Color {
-	return safeColorLookup(ColorNameMenuBackground, currentVariant())
-}
-
-// OverlayBackgroundColor returns the theme's background color for overlays like dialogs.
-//
-// Since: 2.3
-func OverlayBackgroundColor() color.Color {
-	return safeColorLookup(ColorNameOverlayBackground, currentVariant())
-}
-
-// Padding is the standard gap between elements and the border around interface elements.
-func Padding() float32 {
-	return current().Size(SizeNamePadding)
-}
-
-// PlaceHolderColor returns the theme's standard text color.
-func PlaceHolderColor() color.Color {
-	return safeColorLookup(ColorNamePlaceHolder, currentVariant())
-}
-
-// PressedColor returns the color used to overlap tapped features.
-//
-// Since: 2.0
-func PressedColor() color.Color {
-	return safeColorLookup(ColorNamePressed, currentVariant())
-}
-
-// PrimaryColor returns the color used to highlight primary features.
-func PrimaryColor() color.Color {
-	return safeColorLookup(ColorNamePrimary, currentVariant())
-}
-
-// PrimaryColorNamed returns a theme specific color value for a named primary color.
-//
-// Since: 1.4
-func PrimaryColorNamed(name string) color.Color {
-	return primaryColorNamed(name)
-}
-
-// PrimaryColorNames returns a list of the standard primary color options.
-//
-// Since: 1.4
-func PrimaryColorNames() []string {
-	return []string{ColorRed, ColorOrange, ColorYellow, ColorGreen, ColorBlue, ColorPurple, ColorBrown, ColorGray}
-}
-
-// ScrollBarColor returns the color (and translucency) for a scrollBar.
-func ScrollBarColor() color.Color {
-	return safeColorLookup(ColorNameScrollBar, currentVariant())
-}
-
-// ScrollBarSize is the width (or height) of the bars on a ScrollContainer.
-func ScrollBarSize() float32 {
-	return current().Size(SizeNameScrollBar)
-}
-
-// ScrollBarSmallSize is the width (or height) of the minimized bars on a ScrollContainer.
-func ScrollBarSmallSize() float32 {
-	return current().Size(SizeNameScrollBarSmall)
-}
-
-// SelectionColor returns the color for a selected element.
-//
-// Since: 2.1
-func SelectionColor() color.Color {
-	return safeColorLookup(ColorNameSelection, currentVariant())
-}
-
-// SelectionRadiusSize returns the selection highlight radius size.
-//
-// Since: 2.4
-func SelectionRadiusSize() float32 {
-	return current().Size(SizeNameSelectionRadius)
-}
-
-// SeparatorColor returns the color for the separator element.
-//
-// Since: 2.3
-func SeparatorColor() color.Color {
-	return safeColorLookup(ColorNameSeparator, currentVariant())
-}
-
-// SeparatorThicknessSize is the standard thickness of the separator widget.
-//
-// Since: 2.0
-func SeparatorThicknessSize() float32 {
-	return current().Size(SizeNameSeparatorThickness)
-}
-
-// ShadowColor returns the color (and translucency) for shadows used for indicating elevation.
-func ShadowColor() color.Color {
-	return safeColorLookup(ColorNameShadow, currentVariant())
-}
-
-// SuccessColor returns the theme's success foreground color.
-//
-// Since: 2.3
-func SuccessColor() color.Color {
-	return safeColorLookup(ColorNameSuccess, currentVariant())
-}
-
-// TextBoldFont returns the font resource for the bold font style.
-func TextBoldFont() fyne.Resource {
-	return safeFontLookup(fyne.TextStyle{Bold: true})
-}
-
-// TextBoldItalicFont returns the font resource for the bold and italic font style.
-func TextBoldItalicFont() fyne.Resource {
-	return safeFontLookup(fyne.TextStyle{Bold: true, Italic: true})
-}
-
-// TextColor returns the theme's standard text color - this is actually the foreground color since 1.4.
-//
-// Deprecated: Use theme.ForegroundColor() colour instead.
-func TextColor() color.Color {
-	return safeColorLookup(ColorNameForeground, currentVariant())
-}
-
-// TextFont returns the font resource for the regular font style.
-func TextFont() fyne.Resource {
-	return safeFontLookup(fyne.TextStyle{})
-}
-
-// TextHeadingSize returns the text size for header text.
-//
-// Since: 2.1
-func TextHeadingSize() float32 {
-	return current().Size(SizeNameHeadingText)
-}
-
-// TextItalicFont returns the font resource for the italic font style.
-func TextItalicFont() fyne.Resource {
-	return safeFontLookup(fyne.TextStyle{Italic: true})
-}
-
-// TextMonospaceFont returns the font resource for the monospace font face.
-func TextMonospaceFont() fyne.Resource {
-	return safeFontLookup(fyne.TextStyle{Monospace: true})
-}
-
-// TextSize returns the standard text size.
-func TextSize() float32 {
-	return current().Size(SizeNameText)
-}
-
-// TextSubHeadingSize returns the text size for sub-header text.
-//
-// Since: 2.1
-func TextSubHeadingSize() float32 {
-	return current().Size(SizeNameSubHeadingText)
-}
-
-// SymbolFont returns the font resource for the symbol font style.
-//
-// Since: 2.4
-func SymbolFont() fyne.Resource {
-	return safeFontLookup(fyne.TextStyle{Symbol: true})
-}
-
-// WarningColor returns the theme's warning foreground color.
-//
-// Since: 2.3
-func WarningColor() color.Color {
-	return safeColorLookup(ColorNameWarning, currentVariant())
-}
-
 var (
 	defaultTheme fyne.Theme
-
-	errorColor   = color.NRGBA{R: 0xf4, G: 0x43, B: 0x36, A: 0xff}
-	successColor = color.NRGBA{R: 0x43, G: 0xf4, B: 0x36, A: 0xff}
-	warningColor = color.NRGBA{R: 0xff, G: 0x98, B: 0x00, A: 0xff}
 )
 
 type builtinTheme struct {
@@ -887,41 +340,6 @@ func primaryColorNamed(name string) color.NRGBA {
 	// We return the value for ColorBlue for every other value.
 	// There is no need to have it in the switch above.
 	return color.NRGBA{R: 0x29, G: 0x6f, B: 0xf6, A: 0xff}
-}
-
-func safeColorLookup(n fyne.ThemeColorName, v fyne.ThemeVariant) color.Color {
-	col := current().Color(n, v)
-	if col == nil {
-		fyne.LogError("Loaded theme returned nil color", nil)
-		return fallbackColor
-	}
-	return col
-}
-
-func safeFontLookup(s fyne.TextStyle) fyne.Resource {
-	font := current().Font(s)
-	if font != nil {
-		return font
-	}
-	fyne.LogError("Loaded theme returned nil font", nil)
-
-	if s.Monospace {
-		return DefaultTextMonospaceFont()
-	}
-	if s.Bold {
-		if s.Italic {
-			return DefaultTextBoldItalicFont()
-		}
-		return DefaultTextBoldFont()
-	}
-	if s.Italic {
-		return DefaultTextItalicFont()
-	}
-	if s.Symbol {
-		return DefaultSymbolFont()
-	}
-
-	return DefaultTextFont()
 }
 
 func selectionColorNamed(name string) color.NRGBA {
